test: add test coverages
